chore: remove export file
